+++ cx-portal-shared-components/src/components/basic/StaticTemplate/Cards/CardWithoutImage.tsx
@@ -1,5 +1,5 @@
 /********************************************************************************
- * Copyright (c) 2021,2022 BMW Group AG
+ * Copyright (c) 2021,2022 Mercedes-Benz Group AG and BMW Group AG
  * Copyright (c) 2021,2022 Contributors to the Eclipse Foundation
  *
  * See the NOTICE file(s) distributed with this work for additional
@@ -18,36 +18,39 @@
  * SPDX-License-Identifier: Apache-2.0
  ********************************************************************************/
 
-.connector-table-container {
-  margin: 0 10% 70px;
-}
+import { CardDetailsProps } from '../StaticTypes'
+import CardWithText from './CardWithText'
+import { Box } from '@mui/material'
 
-.connector-page-container {
-  margin: 0;
-  padding: 0;
-  width: 100%;
-
-  .picture-with-text-section {
-    margin-top: -200px;
+export default function CardWithoutImage({
+  detail,
+  grid = 3,
+}: {
+  detail: CardDetailsProps
+  grid: number
+}) {
+  const cardContainer = {
+    borderRadius: '16px',
+    display: 'flex',
+    flexDirection: 'column',
+    justifyContent: 'flex-start',
+    marginBottom: '96px',
+    maxWidth: '33%',
   }
-}
-
-.connector-type-selector-container {
-  .dotted-gradient {
-    background-image: linear-gradient(
-      to right,
-      #dadada 40%,
-      rgba(218, 218, 218, 0) 20%
-    );
-    background-position: bottom;
-    background-size: 24px 1.5px;
-    background-repeat: repeat-x;
-    padding: 20px;
-  }
-}
-
-.connector-insert-form {
-  .form-input {
-    margin-bottom: 0px;
-  }
+  return (
+    <Box
+      key={detail.id}
+      sx={{
+        ...cardContainer,
+      }}
+      style={{
+        backgroundColor: detail.backgroundColor,
+        border: '1px solid rgba(15, 113, 203, 1)',
+        padding: '30px',
+        width: `${100 / grid}%`,
+      }}
+    >
+      <CardWithText card={detail} isImage={false} />
+    </Box>
+  )
 }