+++ cx-portal/src/services/AccessService.tsx
@@ -60,6 +60,7 @@ import { EnableIDP } from 'components/overlays/EnableIDP'
 import { EnableIDPSuccess } from 'components/overlays/EnableIDP/EnableIDPSuccess'
 import { DisableIDP } from 'components/overlays/EnableIDP/DisableIDP'
 import { AddusersIDP } from 'components/overlays/AddusersIDP'
+import AddServiceProvider from 'components/overlays/AddServiceProvider'
 
 let pageMap: { [page: string]: IPage }
 let actionMap: { [action: string]: IAction }
@@ -178,6 +179,8 @@ export const getOverlay = (overlay: OverlayState) => {
       )
     case OVERLAYS.SAMPLE_FORM:
       return <SampleForm />
+    case OVERLAYS.ADD_SERVICE_PROVIDER:
+      return <AddServiceProvider id={overlay.id} />
     default:
       return <NotFound />
   }