+++ cx-portal/src/types/Constants.ts
@@ -79,6 +79,8 @@ export enum PAGES {
   INTRODUCTION_APP_PROVIDER = 'companyrolesappprovider',
   INTRODUCTION_SERVICE_PROVIDER = 'companyrolesserviceprovider',
   INTRODUCTION_CONFIRMITY_BODY = 'companyrolesconfirmitybody',
+  USE_CASE = 'usecase',
+  USE_CASE_TRACABILITY = 'usecasetraceablity',
 }
 
 export enum OVERLAYS {
@@ -114,6 +116,7 @@ export enum OVERLAYS {
   APP_DETAILS_OVERLAY = 'app_details_overlay',
   CONFIRM_USER_ACTION = 'confirm_user_action',
   SAMPLE_FORM = 'sample_form',
+  ADD_SERVICE_PROVIDER = 'add_service_provider',
 }
 
 export enum ACTIONS {